+++ src/Components/FlashMessage.vue
@@ -1,56 +1,57 @@
 <script setup lang="ts">
-import {computed} from "vue";
-import {flash, FlashVisuals, defaultFlashVisuals, type FlashContent} from "../Utils/flash.js";
-import {VisualParam, VisualParamDef} from "../Utils/VisualParam.js";
-import {Timer} from "mu-js-utils/lib/Timer.js";
+import {computed, ref} from "vue";
+import { type FlashApi, type FlashVisuals, defaultFlashVisuals } from "../Composables/flash.js";
+import {VisualParam, type VisualParamDef} from "../Utils/VisualParam.js";
+import {Timer} from "mu-js-utils/src/Timer.ts";
 
 const props = defineProps<{
-	message: FlashContent;
-	type?: string;
-	id: number;
+	data: FlashApi;
 	visuals?: VisualParamDef<FlashVisuals>;
 }>();
 
-const timeout = defineModel('timeout', { type: [Number, null] as any, default: 10000 });
-
-const emit = defineEmits(['remove']);
+const timeout = ref<number | null>(props.data.timeout);
 
 const visual = new VisualParam<FlashVisuals>();
-if (props.visuals) visual.addParamDef(props.visuals);
 visual.addParamDef(defaultFlashVisuals);
+if (props.visuals) visual.addParamDef(props.visuals);
 const vis = visual.getAll();
 
 const cssClasses = computed(() => {
-	const resolvedType = flash.typeAlias[props.type] ?? props.type;
+	const resolvedType = props.data.type;
 	return [
-		vis.messageClass,
+		vis.itemClass,
 		(vis.messageTypePrefix ?? '') + resolvedType,
 		timeout.value ? vis.timerClass : null,
 	].flat().filter(Boolean);
 });
 
 const timer = new Timer(timeout.value, true);
-timer.wait().then(() => { emit('remove', props.id); });
+timer.wait().then(() => { props.data.remove(); });
 
 const stopTimer = () => {
 	timer.stop();
 	timeout.value = null;
-}
+};
+
 const messageType = computed(() => {
-  if (typeof props.message === 'string') return 'text';
-  else if (props.message.html) return 'html';
+  const msg = props.data.message as any;
+  if (typeof msg === 'string') return 'text';
+  else if (msg.html) return 'html';
   else return 'component';
 });
 </script>
 
+
 <template>
 <div :class="cssClasses" @click="stopTimer" :style="{'--flash-close-timeout': timeout + 'ms'}">
-  <template v-if="messageType === 'text'">
-	  {{ message }}
-  </template>
-  <template  v-if="messageType === 'html'" v-html="message.html" />
-  <component v-if="messageType === 'component'" :is="props.message.component" v-bind="props.message.componentProps" />
-	<button type="button" :class="vis.closeBtnClass" @click="$emit('remove', props.id)" aria-label="Close"></button>
+  <div :class="vis.messageClass">
+    <template v-if="messageType === 'text'">
+      {{ data.message }}
+    </template>
+    <div v-if="messageType === 'html'" v-html="(data.message as any).html" />
+    <component v-if="messageType === 'component'" :is="(data.message as any).component" v-bind="(data.message as any).componentProps" />
+  </div>
+	<button type="button" :class="vis.closeBtnClass" @click.stop="data.remove()" aria-label="Close">{{ vis.closeBtnText }}</button>
 </div>
 </template>