+++ src/Utils/types.ts
@@ -1,0 +1,8 @@
+import { Component, ConcreteComponent } from "vue";
+
+export type ElementContent = string
+    | { html: string }
+    | {
+    component: Component | ConcreteComponent;
+    componentProps?: Record<string, any>;
+};